pyproject.toml
[tool.poetry]
name = "ai-micro-api-doc"
version = "1.0.0"
description = "ドキュメント処理サービス - OCR、レイアウト解析、チャンク分割"
authors = ["AI Micro Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pydantic = "^2.9.0"
pydantic-settings = "^2.5.0"
sqlalchemy = "^2.0.35"
psycopg2-binary = "^2.9.9"
redis = "^5.2.0"
httpx = "^0.27.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
celery = "^5.3.6"

# Document Processing (from api-admin)
# docling 2.x includes docling-parse 2.x+ with UTF-8 handling fixes
docling = "^2.0.0"
pypdfium2 = "^4.25.0"
pillow = "^10.0.0"

# OCR
easyocr = "^1.7.2"
pytesseract = "^0.3.13"

# Embedding
langchain = "^0.3.0"
langchain-community = "^0.3.0"
langchain-postgres = "^0.0.12"
sentence-transformers = "^3.2.0"
pgvector = "^0.2.5"

# Japanese tokenization
mecab-python3 = "^1.0.10"
unidic-lite = "^1.0.8"

# ML
torch = "^2.0.0"
numpy = "^1.26.0"

# Utilities
python-multipart = "^0.0.12"
psutil = "^5.9.0"
async-timeout = "^4.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^5.0.0"
ruff = "^0.6.0"
mypy = "^1.11.0"
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "D", "UP", "B", "C4", "SIM"]
ignore = ["D100", "D104", "D107", "D203", "D213"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
